python: add read_physical
